% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/above.R
\name{above}
\alias{above}
\title{Calculate sum of observations above distance value}
\usage{
above(df, run, dist_value = 0.9)
}
\arguments{
\item{df}{A data frame to run the analysis on}

\item{run}{The number of the specific iteration run}

\item{dist_value}{The distance value to calculate from}
}
\description{
Calculates the sum of observations above dist_value and calculates the percentage of observed values per group
}
