% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/below.R
\name{below}
\alias{below}
\title{Calculate sum of observations below distance value}
\usage{
below(df, run, dist_value = 0.1)
}
\arguments{
\item{df}{A data frame to run the analysis on}

\item{run}{The number of the specific iteration run}

\item{dist_value}{The distance value to calculate from}
}
\description{
Calculates the sum of observations below dist_value and calculates the percentage of observed values per group
}
