% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_paste.R
\name{func_paste}
\alias{func_paste}
\title{Paste unique values in column together}
\usage{
func_paste(x)
}
\arguments{
\item{x}{A column name}
}
\description{
Collapses column based on unique values in given column
}
