% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_samples.R
\name{iterate_samples}
\alias{iterate_samples}
\title{Iterate samples}
\usage{
iterate_samples(df, run, method = "mean", dist_value1, dist_value2)
}
\arguments{
\item{df}{A data frame to run the analysis on}

\item{run}{The number of the specific iteration run}

\item{method}{The summarizing method used, either "mean", "below", "above", "within", "without", "median", or "quantile"}

\item{dist_value1}{The first distance value to calculate from, usually the lower value}

\item{dist_value2}{The second distance value to calculate from, usually the higher value}
}
\description{
Specify iteration method and iterate over samples
}
