% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_calc.R
\name{mean_calc}
\alias{mean_calc}
\title{Calculate mean value of distances}
\usage{
mean_calc(df, run)
}
\arguments{
\item{df}{A data frame to run the analysis on}

\item{run}{The number of the specific iteration run}
}
\description{
Calculates the mean of distances per group
}
