% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_calc.R
\name{median_calc}
\alias{median_calc}
\title{Calculate median value of distances}
\usage{
median_calc(df, run)
}
\arguments{
\item{df}{A data frame to run the analysis on}

\item{run}{The number of the specific iteration run}
}
\description{
Calculates the median of observations for each group
}
