% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist.R
\name{plot_dist}
\alias{plot_dist}
\title{Plot results from iterations}
\usage{
plot_dist(df)
}
\arguments{
\item{df}{The data frame with results from the "iterate" function}
}
\description{
Plots the distribution of the expected values relative to the observed.
}
