% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_calc.R
\name{quantile_calc}
\alias{quantile_calc}
\title{Calculate quantile values of distances}
\usage{
quantile_calc(df, run)
}
\arguments{
\item{df}{A data frame to run the analysis on}

\item{run}{The number of the specific iteration run}
}
\description{
Calculates the quantile values of observations for each group
}
