% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within.R
\name{within}
\alias{within}
\title{Calculate sum of observations between distance 1 and distance 2}
\usage{
within(df, run, dist_value1 = 0.1, dist_value2 = 0.8)
}
\arguments{
\item{df}{A data frame to run the analysis on}

\item{run}{The number of the specific iteration run}

\item{dist_value1}{The first distance value to calculate from, usually the lower value}

\item{dist_value2}{The second distance value to calculate from, usually the higher value}
}
\description{
Calculates the sum of observations between dist_value1 and dist_value2, and calculates the percentage of observed values per group
}
