% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/without.R
\name{without}
\alias{without}
\title{Calculate sum of observations below distance 1 and above distance 2}
\usage{
without(df, run, dist_value1 = 0.1, dist_value2 = 0.8)
}
\arguments{
\item{df}{A data frame to run the analysis on}

\item{run}{The number of the specific iteration run}

\item{dist_value1}{The first distance value to calculate from, usually the lower value}

\item{dist_value2}{The second distance value to calculate from, usually the higher value}
}
\description{
Calculates the sum of observations below dist_value1 and above dist_value2, and calculates the percentage of observed values per group
}
